requirements.txt
# Core dependencies
streamlit>=1.28.0
pandas>=1.5.0
matplotlib>=3.6.0
scikit-learn>=1.2.0
Pillow>=9.0.0

# Data processing
numpy>=1.21.0
scipy>=1.9.0

# Machine Learning
joblib>=1.2.0

# OCR and Image Processing
google-cloud-vision>=3.4.0
python-dotenv>=1.0.0

# Voice Recognition
SpeechRecognition>=3.10.0
pyaudio>=0.2.11

# Cloud Services
gspread>=5.7.0
oauth2client>=4.1.3
openai>=1.0.0

# Optional: Firebase (may require additional setup)
# pyrebase>=4.0.0

# Additional utilities
altair>=4.2.0
plotly>=5.15.0
seaborn>=0.11.0

# Enhanced visualization
kaleido>=0.2.1

# Export tools
xlsxwriter>=3.0.0
fpdf>=1.7.2
